+++ apps/courses/utils/exceptions.py
